refactor conventions submodule
