--- sofar/conventions/GeneralTF-E_1.0.json
@@ -1,277 +1,0 @@
-{
-    "GLOBAL:Conventions": {
-        "default": "SOFA",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:Version": {
-        "default": "2.0",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:SOFAConventions": {
-        "default": "GeneralTF-E",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": "This conventions stores TFs depending in the Emiiter for general purposes, i.e., only the mandatory, SOFA general metadata are pre-defined. This convention is based on GeneralTF"
-    },
-    "GLOBAL:SOFAConventionsVersion": {
-        "default": "1.0",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:APIName": {
-        "default": "",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:APIVersion": {
-        "default": "",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:ApplicationName": {
-        "default": "",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:ApplicationVersion": {
-        "default": "",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:AuthorContact": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:Comment": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:DataType": {
-        "default": "TF-E",
-        "flags": "rm",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": "We store frequency-dependent data depending on the emitter here"
-    },
-    "GLOBAL:History": {
-        "default": "",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:License": {
-        "default": "No license provided, ask the author for permission",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:Organization": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:References": {
-        "default": "",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:RoomType": {
-        "default": "free field",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": "The room information can be arbitrary"
-    },
-    "GLOBAL:Origin": {
-        "default": "",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:DateCreated": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:DateModified": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "GLOBAL:Title": {
-        "default": "",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "ListenerPosition": {
-        "default": [
-            0,
-            0,
-            0
-        ],
-        "flags": "m",
-        "dimensions": "IC, MC",
-        "type": "double",
-        "comment": ""
-    },
-    "ListenerPosition:Type": {
-        "default": "cartesian",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "ListenerPosition:Units": {
-        "default": "metre",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "ReceiverPosition": {
-        "default": [
-            0,
-            0,
-            0
-        ],
-        "flags": "m",
-        "dimensions": "IC, RC, RCM",
-        "type": "double",
-        "comment": ""
-    },
-    "ReceiverPosition:Type": {
-        "default": "cartesian",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "ReceiverPosition:Units": {
-        "default": "metre",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "SourcePosition": {
-        "default": [
-            0,
-            0,
-            1
-        ],
-        "flags": "m",
-        "dimensions": "IC, MC",
-        "type": "double",
-        "comment": "In order to store different directions/positions around the listener, SourcePosition is assumed to vary"
-    },
-    "SourcePosition:Type": {
-        "default": "spherical",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "SourcePosition:Units": {
-        "default": "degree, degree, metre",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "EmitterPosition": {
-        "default": [
-            0,
-            0,
-            0
-        ],
-        "flags": "m",
-        "dimensions": "IC, EC, ECM",
-        "type": "double",
-        "comment": ""
-    },
-    "EmitterPosition:Type": {
-        "default": "cartesian",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "EmitterPosition:Units": {
-        "default": "metre",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "N": {
-        "default": 0,
-        "flags": "m",
-        "dimensions": "N",
-        "type": "double",
-        "comment": "Frequency values"
-    },
-    "N:LongName": {
-        "default": "frequency",
-        "flags": null,
-        "dimensions": null,
-        "type": "attribute",
-        "comment": ""
-    },
-    "N:Units": {
-        "default": "hertz",
-        "flags": "m",
-        "dimensions": null,
-        "type": "attribute",
-        "comment": "Unit of the values given in N"
-    },
-    "Data.Real": {
-        "default": 0,
-        "flags": "m",
-        "dimensions": "mrne",
-        "type": "double",
-        "comment": "The real part of the complex spectrum"
-    },
-    "Data.Imag": {
-        "default": 0,
-        "flags": "m",
-        "dimensions": "MRNE",
-        "type": "double",
-        "comment": "The imaginary part of the complex spectrum"
-    }
-}
