--- sofar/conventions/source/GeneralTF_1.0.csv
@@ -1,38 +1,0 @@
-Name	Default	Flags	Dimensions	Type	Comment
-GLOBAL:Conventions	SOFA	rm	  	attribute
-GLOBAL:Version	1.0	rm	 	attribute
-GLOBAL:SOFAConventions	GeneralTF	rm	  	attribute	This conventions stores TFs for general purposes, i.e., only the mandatory, SOFA general metadata are pre-defined. This convention is based on GeneralFIR.
-GLOBAL:SOFAConventionsVersion	1.0	rm	  	attribute
-GLOBAL:APIName		rm	  	attribute
-GLOBAL:APIVersion		rm	  	attribute
-GLOBAL:ApplicationName			  	attribute
-GLOBAL:ApplicationVersion			  	attribute
-GLOBAL:AuthorContact		m	  	attribute
-GLOBAL:Comment		m	 	attribute
-GLOBAL:DataType	TF	rm	  	attribute	We store frequency-dependent data here
-GLOBAL:History			 	attribute
-GLOBAL:License	No license provided, ask the author for permission	m	  	attribute
-GLOBAL:Organization		m	  	attribute
-GLOBAL:References				attribute
-GLOBAL:RoomType	free field	m	  	attribute	The room information can be arbitrary
-GLOBAL:Origin				attribute
-GLOBAL:DateCreated		m	 	attribute
-GLOBAL:DateModified		m	 	attribute
-GLOBAL:Title		m		attribute
-ListenerPosition	[0 0 0] 	m	IC, MC	double
-ListenerPosition:Type	cartesian	m	  	attribute
-ListenerPosition:Units	metre	m	  	attribute
-ReceiverPosition	[0 0 0]	m	rCI, rCM	double
-ReceiverPosition:Type	cartesian	m	  	attribute
-ReceiverPosition:Units	metre	m	  	attribute
-SourcePosition	[0 0 1]	m	IC, MC	double	In order to store different directions/positions around the listener, SourcePosition is assumed to vary
-SourcePosition:Type	spherical	m	  	attribute
-SourcePosition:Units	degree, degree, metre	m	  	attribute
-EmitterPosition	[0 0 0]	m	eCI, eCM	double
-EmitterPosition:Type	cartesian	m	  	attribute
-EmitterPosition:Units	metre	m	  	attribute
-Data.Real	0	m	mRn	double	The real part of the complex spectrum
-Data.Imag	0	m	MRN	double	The imaginary part of the complex spectrum
-N	0	m	N	double	Frequency values
-N:LongName	frequency	m		attribute
-N:Units	hertz	m		attribute	Unit of the values given in N
