--- sofar/conventions/source/MultiSpeakerBRIR_0.3.csv
@@ -1,48 +1,0 @@
-Name	Default	Flags	Dimensions	Type	Comment
-GLOBAL:Conventions	SOFA	rm	  	attribute
-GLOBAL:Version	1.0	rm	 	attribute
-GLOBAL:SOFAConventions	MultiSpeakerBRIR	rm	  	attribute	This convention is for BRIRs recorded in reverberant conditions from multiple loudspeaker sources at a number of listener orientations.
-GLOBAL:SOFAConventionsVersion	0.3	rm	  	attribute
-GLOBAL:APIName		rm	  	attribute
-GLOBAL:APIVersion		rm	  	attribute
-GLOBAL:ApplicationName			  	attribute
-GLOBAL:ApplicationVersion			  	attribute
-GLOBAL:AuthorContact		m	  	attribute
-GLOBAL:Comment		m	 	attribute
-GLOBAL:DataType	FIRE	rm	  	attribute	We use FIR datatype which in addition depends on Emitters (E)
-GLOBAL:History			 	attribute
-GLOBAL:License	No license provided, ask the author for permission	m	  	attribute
-GLOBAL:Organization		m	  	attribute
-GLOBAL:References				attribute
-GLOBAL:RoomType	reverberant	m	  	attribute
-GLOBAL:Origin				attribute
-GLOBAL:DateCreated		m	 	attribute
-GLOBAL:DateModified		m	 	attribute
-GLOBAL:Title		m		attribute
-ListenerPosition	[0 0 0] 	m	IC, MC	double
-ListenerPosition:Type	cartesian	m	  	attribute
-ListenerPosition:Units	metre	m	  	attribute
-ReceiverPosition	[0 0.09 0; 0 -0.09 0]	m	rCI, rCM	double
-ReceiverPosition:Type	cartesian	m	  	attribute
-ReceiverPosition:Units	metre	m	  	attribute
-SourcePosition	[0 0 1]	m	IC, MC	double
-SourcePosition:Type	spherical	m	  	attribute
-SourcePosition:Units	degree, degree, metre	m	  	attribute
-EmitterPosition	[0 0 0]	m	eCI, eCM	double	Each speaker is represented as an emitter. Use EmitterPosition to represent the position of a particular speaker. Size of EmitterPosition determines E
-EmitterPosition:Type	cartesian	m	  	attribute
-EmitterPosition:Units	metre	m	  	attribute
-GLOBAL:DatabaseName		m	  	attribute	name of the database to which these data belong
-GLOBAL:ListenerShortName		m	  	attribute	ID of the subject from the database
-GLOBAL:RoomDescription				attribute	narrative description of the room
-ListenerUp	[0 0 1]	m	IC, MC	double
-ListenerView	[1 0 0]	m	IC, MC	double
-ListenerView:Type	cartesian	m	  	attribute
-ListenerView:Units	metre	m	  	attribute
-EmitterUp	[0 0 1]		ECI, ECM	double	When EmitterUp provided, EmitterView must be provided as well
-EmitterView	[1 0 0]		ECI, ECM	double	When EmitterView provided, EmitterUp must be provided as well
-EmitterView:Type	cartesian		  	attribute
-EmitterView:Units	metre		  	attribute
-Data.IR	[1 1]	m	mREn	double
-Data.SamplingRate	48000	m	I	double
-Data.SamplingRate:Units	hertz	m		attribute
-Data.Delay	[0 0]	m	IRE, MRE	double
