--- sofar/conventions/source/SingleRoomMIMOSRIR_1.0.csv
@@ -1,68 +1,0 @@
-Name	Default	Flags	Dimensions	Type	Comment
-GLOBAL:Conventions	SOFA	rm	  	attribute
-GLOBAL:Version	2.0	rm	 	attribute
-GLOBAL:SOFAConventions	SingleRoomMIMOSRIR	rm	  	attribute	Single-room multiple-input multiple-output spatial room impulse responses, depending on Emitters
-GLOBAL:SOFAConventionsVersion	1.0	rm	  	attribute
-GLOBAL:DataType	FIR-E	rm	  	attribute	Shall be FIR-E
-GLOBAL:RoomType	shoebox	m	  	attribute	Shall be 'shoebox' or 'dae'
-GLOBAL:Title		m		attribute
-GLOBAL:DateCreated		m	 	attribute
-GLOBAL:DateModified		m	 	attribute
-GLOBAL:APIName		rm	  	attribute
-GLOBAL:APIVersion		rm	  	attribute
-GLOBAL:AuthorContact		m	  	attribute
-GLOBAL:Organization		m	  	attribute
-GLOBAL:License	No license provided, ask the author for permission	m	  	attribute
-GLOBAL:ApplicationName			  	attribute
-GLOBAL:ApplicationVersion			  	attribute
-GLOBAL:Comment			 	attribute
-GLOBAL:RoomDescription				attribute	narrative description of the room
-GLOBAL:History			 	attribute
-GLOBAL:References				attribute
-GLOBAL:Origin				attribute
-GLOBAL:ListenerShortName				attribute
-GLOBAL:ListenerDescription				attribute
-ListenerPosition	[0 0 0] 	m	MC	double
-ListenerPosition:Type	cartesian	m	  	attribute
-ListenerPosition:Units	metre	m	  	attribute
-ListenerView	[1 0 0]	m	IC, MC	double
-ListenerUp	[0 0 1]	m	IC, MC	double
-ListenerView:Type	cartesian	m		attribute
-ListenerView:Units	metre	m		attribute
-GLOBAL:ReceiverShortName				attribute
-GLOBAL:ReceiverDescription				attribute
-ReceiverPosition	[0 0 0]	m	IC, RCI, RCM	double
-ReceiverPosition:Type	spherical	m	  	attribute	Can be of any type enabling both spatially discrete and spatially continuous representations.
-ReceiverPosition:Units	degree, degree, metre	m	  	attribute
-ReceiverView	[1 0 0]		RCI, RCM	double
-ReceiverUp	[0 0 1]		RCI, RCM	double
-ReceiverView:Type	cartesian			attribute
-ReceiverView:Units	metre			attribute
-GLOBAL:SourceShortName				attribute
-GLOBAL:SourceDescription				attribute
-SourcePosition	[0 0 1]	m	MC	double
-SourcePosition:Type	cartesian	m	  	attribute
-SourcePosition:Units	metre	m	  	attribute
-SourceView	[1 0 0]	m	IC, MC	double
-SourceUp	[0 0 1]	m	IC, MC	double
-SourceView:Type	cartesian	m		attribute
-SourceView:Units	metre	m		attribute
-GLOBAL:EmitterShortName				attribute
-GLOBAL:EmitterDescription				attribute
-EmitterPosition	[0 0 0]	m	IC, ECI, ECM	double	Can be of any type enabling both spatially discrete and spatially continuous representations.
-EmitterPosition:Type	spherical	m	  	attribute
-EmitterPosition:Units	degree, degree, metre	m	  	attribute
-EmitterView	[1 0 0]		ECI, ECM	double
-EmitterUp	[0 0 1]		ECI, ECM	double
-EmitterView:Type	cartesian			attribute
-EmitterView:Units	metre			attribute
-Data.IR	0	m	mrne	double	Impulse responses
-Data.SamplingRate	48000	m	I, M	double	Sampling rate of the samples in Data.IR and Data.Delay
-Data.SamplingRate:Units	hertz	m		attribute	Unit of the sampling rate
-Data.Delay	0	m	IRI, MRI, MRE	double	Additional delay of each IR (in samples)
-RoomCornerA	[0 0 0]	m	IC, MC	double
-RoomCornerB	[1 2 3]	m	IC, MC	double
-RoomCorners	0		II	double	The value of this attribute is to be ignored. It only exist to for RoomCorners:Type and RoomCorners:Units
-RoomCorners:Type	cartesian	m		attribute
-RoomCorners:Units	metre	m		attribute
-GLOBAL:DatabaseName		m		attribute	Name of the database. Used for classification of the data.
